generateBlock extracted from ComponentChartsPage, and Stickers are added.
