+++ src/components/base/PageBase.js
@@ -22,6 +22,9 @@ import TvIcon from '@material-ui/icons/Tv';
 
 import { Board } from '@stickyboard/core';
 
+import Stickers from 'components/stickers';
+import generateBlock2 from 'components/stickers/generateBlock';
+
 import ApiManager from 'network/ApiManager';
 import StatusCode from 'network/StatusCode';
 import CookieManager from 'network/CookieManager';
@@ -166,7 +169,25 @@ class PageBase extends React.Component {
                     }}
                     onSaveLayout={this.onSaveLayout}>
                     {blocks.map((block, index) => {
-                        return generateBlock(block);
+                        if (generateBlock) {
+                            return generateBlock(block);
+                        } else {
+                            return generateBlock2(theme.colors.colorArray, block)
+
+                            // const Sticker = Stickers[block.i];
+                            //
+                            // if (
+                            //     Sticker &&
+                            //     typeof Sticker.Component === 'function'
+                            // ) {
+                            //     return (
+                            //         <Sticker.Component
+                            //             key={block.i}
+                            //             colors={theme.colors.colorArray}
+                            //         />
+                            //     );
+                            // }
+                        }
                     })}
                 </Board>
 
@@ -212,7 +233,7 @@ PageBase.propTypes = {
     classes: PropTypes.object.isRequired,
     theme: PropTypes.object.isRequired,
     // Layout
-    generateBlock: PropTypes.func.isRequired,
+    // generateBlock: PropTypes.func.isRequired,
     initialLayout: PropTypes.object.isRequired,
     initialBlocks: PropTypes.array.isRequired,
 };
