+++ src/components/stickers/RadarChartSticker.js
@@ -1,0 +1,41 @@
+import React from 'react';
+import { Sticker } from '@stickyboard/core';
+import { RadarChart } from '@stickyboard/recharts';
+
+const radarChartData = [
+    { subject: 'Math', A: 120, B: 70, fullMark: 150 },
+    { subject: 'Chinese', A: 98, B: 60, fullMark: 150 },
+    { subject: 'English', A: 86, B: 130, fullMark: 150 },
+    { subject: 'Geography', A: 99, B: 100, fullMark: 150 },
+    { subject: 'Physics', A: 85, B: 90, fullMark: 150 },
+    { subject: 'History', A: 65, B: 85, fullMark: 150 },
+];
+
+function RadarChartSticker(props) {
+    const { colors } = props;
+
+    return (
+        <Sticker>
+            <RadarChart
+                data={radarChartData}
+                polarAngleAxisKey={'subject'}
+                radarAttrArray={[
+                    {
+                        name: 'Mike',
+                        dataKey: 'A',
+                        stroke: '#ffed00',
+                        fill: colors[4],
+                    },
+                    {
+                        name: 'Lily',
+                        dataKey: 'B',
+                        stroke: '#66d522',
+                        fill: colors[5],
+                    },
+                ]}
+            />
+        </Sticker>
+    );
+}
+
+export default RadarChartSticker;
